Reorganize abc finder app into separate files.
